backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
python-dotenv==1.0.0
web3==6.0.0
pandas==2.0.0
numpy==1.24.3
scikit-learn==1.2.2
lightgbm==3.3.5
shap==0.41.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.9
alembic==1.10.3
pytest==7.3.1
httpx==0.23.3
